refactor(model): drop redundant guard + SF-1 logging, isolate download I/O root
Three changes the coverage gate surfaced as untestable production lines:

- Drop the ChunkIdLengthMismatch guard in index_corpus. rurico c862123's
  ChunkedEmbedding::try_new now generates chunk_ids from chunks (equal length)
  and the fields are pub(crate), so the length mismatch is unreachable — the
  guard re-checks an invariant the tightened contract enforces upstream.
- Revert the SF-1 cache-error logging: out of scope for the bump (the dropped
  arm is pre-existing) and only reachable on a real cache IO error, so it
  cannot be unit-tested.
- Move download_and_verify_model into model/download.rs: the I/O composition
  root (real network download + on-device MLX probe). The logic it wires is
  covered via try_download_and_verify_with_fns; isolating it lets the diff
  coverage gate exclude only the irreducible I/O, not model.rs's tested code.
