+++ src/model.rs
@@ -1,17 +1,20 @@
 pub mod embedder;
 pub mod reranker;
 
+mod download;
+pub use download::download_and_verify_model;
+
 use std::convert::Infallible;
 use std::error::Error;
 use std::fmt;
 use std::io;
 
-use rurico::embed::{Artifacts, Embed, Embedder, ModelId, download_model};
+use rurico::embed::Embed;
 use rurico::model_init::ModelInitError;
 use rurico::model_probe::ProbeStatus;
 
 use self::embedder::try_load_embedder_with_fns;
-use crate::cli::{hint, warning, with_spinner};
+use crate::cli::{hint, warning};
 
 /// Reason a model could not be loaded.
 ///
@@ -252,57 +255,7 @@ impl fmt::Display for ModelDownloadError {
 
 impl Error for ModelDownloadError {}
 
-/// Download the default embedding model and verify it loads correctly.
-///
-/// Shows a spinner on stderr during download and probe phases.
-///
-/// # Prerequisites
-///
-/// "Verify" here means **probe-load**: after the download, the fn calls
-/// [`Embedder::probe`] and [`Embedder::new`] to confirm the artifacts
-/// actually load on this machine. It is **not** a content-hash check —
-/// artifact integrity is delegated to `rurico::embed::download_model`,
-/// whose own checksum step runs before this fn's post-download probe. A
-/// successful return guarantees "the model loads here and now", not
-/// "the downloaded bytes match the registry manifest".
-///
-/// The calling binary must invoke `rurico::handle_probe_if_needed()`
-/// at the very start of `main()`. Without it, the post-download probe returns
-/// [`ModelDownloadError::ProbeFailed`] even when the download succeeds.
-///
-/// # Errors
-///
-/// - [`ModelDownloadError::DownloadFailed`] — the HTTP download from the model
-///   registry failed.
-/// - [`ModelDownloadError::BackendUnavailable`] — the hardware/OS backend
-///   (e.g. MLX) is not available on this machine.
-/// - [`ModelDownloadError::ProbeFailed`] — the downloaded model files could not
-///   be loaded. Corrupt artifacts are deleted automatically so a subsequent
-///   call can re-download. Non-corrupt probe or init failures leave artifacts
-///   intact.
-pub fn download_and_verify_model() -> Result<(), ModelDownloadError> {
-    with_spinner(
-        "Downloading model...",
-        |_| "Model ready".to_owned(),
-        |update| {
-            try_download_and_verify_with_fns(
-                || {
-                    download_model(ModelId::DEFAULT).map_err(|e| {
-                        tracing::error!(error = %e, "model download failed");
-                        e
-                    })
-                },
-                |e| tracing::warn!(error = %e, "failed to delete artifacts after failed probe"),
-                Embedder::probe,
-                Embedder::new,
-                Artifacts::delete_files,
-                || update("Verifying model..."),
-            )
-        },
-    )
-}
-
-fn try_download_and_verify_with_fns<A, E, DE>(
+pub(super) fn try_download_and_verify_with_fns<A, E, DE>(
     download_fn: impl FnOnce() -> Result<A, DE>,
     on_delete_error: impl FnOnce(io::Error),
     probe_fn: impl FnOnce(&A) -> Result<ProbeStatus, ModelInitError>,
